[#8] Add function that performs common api setup and refactor lsr demo api to use it
