+++ demo_api/lsr/Dockerfile
@@ -1,11 +1,11 @@
 FROM python:3.8-buster
 
-COPY . /lsr/
+COPY . /demo_api
 
-WORKDIR /lsr
+WORKDIR /demo_api/lsr
 
 RUN pip install -r requirements.txt
 
 RUN python -m download_pretrained
 
-CMD gunicorn --bind 0.0.0.0:8000 wsgi:app --timeout 180
+CMD PYTHONPATH=../ gunicorn --bind 0.0.0.0:8000 api:app --timeout 180
