--- demo_api/lsr/wsgi.py
@@ -1,4 +1,0 @@
-from model_api import app
-
-if __name__ == '__main__':
-    app.run()
